mitosheet: cleanup schedule code
